+++ CHANGELOG.md
@@ -3,6 +3,9 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+- Letter.stream() and Letter.stream(startingLetter) added.
+
 ### Changed
 - Letter.iterator() now loops through all the letters indefinitely.